Refactored to put the CA classes in their own folder.
