+++ conway_module/lib/templates/automata/LinearCellularAutomaton.js
@@ -1,27 +1,18 @@
-const DrawingTemplate = require('./DrawingTemplate.js');
+const DrawingTemplate = require('./../DrawingTemplate.js');
 
 class LinearCellularAutomaton extends DrawingTemplate {
 	constructor(config) {
 		super();
 		this.config = config;
+		this.rulesSet = null;
 	}
 
 	generateCells(x, y) {
 		let cells = this.makeCellsFrom2DArray(this.pattern());
 		return cells;
 	}
-}
-
-class WolframRule90 extends LinearCellularAutomaton {
-	constructor(config) {
-		super(config);
-		this.rulesSet = [0, 1, 0, 1, 1, 0, 1, 0]; //Stored in reverse order from Wolfram's Specificiation to enable binary lookup.
-	}
 
 	pattern() {
-		console.log(
-			`Landscape width: ${this.config.landscape.width} height: ${this.config.landscape.height}`
-		);
 		let height = Math.floor(this.config.landscape.height);
 		let width = Math.floor(this.config.landscape.width);
 
@@ -35,7 +26,7 @@ class WolframRule90 extends LinearCellularAutomaton {
 
 		//Generate the next row based on the current row.
 		let neighborhood;
-		for (let row = 0; row < height - 2; row++) {
+		for (let row = 0; row < height; row++) {
 			ca[row + 1] = Array(width);
 			for (let col = 0; col < width; col++) {
 				neighborhood = this.findNeighborHood(ca[row], col);
@@ -61,9 +52,21 @@ class WolframRule90 extends LinearCellularAutomaton {
 	}
 
 	evaluateRules(neighborhood) {
-		let rule = parseInt(neighborhood.join(''), 2);
-		return this.rulesSet[rule];
+		let ruleIndex = parseInt(neighborhood.join(''), 2);
+		return this.getRule(ruleIndex);
+	}
+
+	getRule(ruleIndex) {
+		return this.rulesSet[ruleIndex];
+	}
+
+	/**
+	 * Stores the rule set in reverse order from Wolfram's Specificiation to enable binary lookup.
+	 * @param {number[]} ruleSet
+	 */
+	setRuleSet(ruleSet) {
+		this.rulesSet = ruleSet;
 	}
 }
 
-module.exports = WolframRule90;
+module.exports = LinearCellularAutomaton;
